Squashed commit of changes dealing with string handling.
The Dimensions/ Shapes/ and Metadata/ groups and the
RetrievalResults/aerosol_model, RetrievedStateVector/state_vector_names data fields
are now taken from the first retrieval file and written to the aggregated file.
RetrievalResults/surface_type is taken from each file and combined into an array
of fixed length strings (S19) of size [sounding]. It may become an issue that
this is not [sounding, fooprint], but we will need to cross that bridge when
we come to it.

I also noticed a couple of unrelated possible bugs that will need to be checked
in future branches:

1) There were entries in indexes where there should not have been (no retrieval file
   was produced for those indexes). Example fields that exibited the issue:
       RetrievalHeader/*
       RetrievalResults/apriori_o2_column,
       RetrievalResults/brdf_anisotropy_ch4

2) There was an OOM error when I tried to open some fields in HDFView. Example fields:
       SpectralParameters/measured_radiance
       SpectralParameters/measured_radiance_uncert
       SpectralParameters/wavelength